SKIN CARE/requirements.txt
Flask==3.0.2
Werkzeug==3.0.1
torch==2.2.2
torchvision==0.17.2
pillow==10.3.0
